wifi: mac80211_hwsim: add back erroneously removed cast
commit 58b6259d820d63c2adf1c7541b54cce5a2ae6073 upstream.

The robots report that we're now casting to a differently
sized integer, which is correct, and the previous patch
had erroneously removed it.

Reported-by: kernel test robot <lkp@intel.com>
Fixes: 4ee186fa7e40 ("wifi: mac80211_hwsim: fix race condition in pending packet")
Cc: Jeongik Cha <jeongik@google.com>

+++ drivers/net/wireless/mac80211_hwsim.c
@@ -3155,7 +3155,8 @@ static int hwsim_tx_info_frame_received_nl(struct sk_buff *skb_2,
 
 		txi = IEEE80211_SKB_CB(skb);
 
-		skb_cookie = (u64)txi->rate_driver_data[0];
+		skb_cookie = (u64)(uintptr_t)txi->rate_driver_data[0];
+
 
 
 		if (skb_cookie == ret_skb_cookie) {
